Populate CSI config with migration datasore url
