+++ assets/csi_cloud_config.ini
@@ -6,4 +6,5 @@ cluster-id = "${CLUSTER_ID}"
 [VirtualCenter "${VCENTER}"]
 insecure-flag = "true"
 datacenters = "${DATACENTERS}"
+migration-datastore-url = "${MIGRATION_DATASTORE_URL}"
