+++ pkg/operator/storageclasscontroller/storageclasscontroller.go
@@ -63,7 +63,7 @@ func NewStorageClassController(
 		kubeClient:           kubeClient,
 		operatorClient:       operatorClient,
 		recorder:             recorder,
-		makeStoragePolicyAPI: newStoragePolicyAPI,
+		makeStoragePolicyAPI: NewStoragePolicyAPI,
 		scStateEvaluator:     evaluator,
 	}
